--- app/src/main/java/ru/netology/nmedia/activity/ActivityNewPost.kt
@@ -1,35 +1,0 @@
-package ru.netology.nmedia.activity
-
-import android.app.Activity
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import ru.netology.nmedia.databinding.ActivityNewPostBinding
-
-class ActivityNewPost : AppCompatActivity() {
-    lateinit var binding: ActivityNewPostBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityNewPostBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        binding.apply {
-
-            edit.setText(intent.extras?.getString(Intent.EXTRA_TITLE))
-
-            edit.requestFocus()
-
-            ok.setOnClickListener {
-                val intent = Intent()
-                if (binding.edit.text.isNullOrBlank()) {
-                    setResult(Activity.RESULT_CANCELED, intent)
-                } else {
-                    val content = binding.edit.text.toString()
-                    intent.putExtra(Intent.EXTRA_TEXT, content)
-                    setResult(Activity.RESULT_OK, intent)
-                }
-                finish()
-            }
-        }
-    }
-}
