+++ app/src/main/java/ru/netology/nmedia/screens/AppActivity.kt
@@ -1,0 +1,13 @@
+package ru.netology.nmedia.screens
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import ru.netology.nmedia.databinding.ActivityAppBinding
+
+class AppActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        val binding = ActivityAppBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+    }
+}
