+++ app/src/main/java/ru/netology/nmedia/screens/FeedFragment.kt
@@ -1,32 +1,35 @@
-package ru.netology.nmedia.activity
+package ru.netology.nmedia.screens
 
 import android.content.Intent
 import android.net.Uri
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.Toast
-import androidx.activity.viewModels
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.activityViewModels
+import androidx.navigation.fragment.findNavController
 import ru.netology.nmedia.R
 import ru.netology.nmedia.adapter.OnInteractionListener
 import ru.netology.nmedia.adapter.PostAdapter
-import ru.netology.nmedia.databinding.ActivityMainBinding
+import ru.netology.nmedia.databinding.FragmentFeedBinding
 import ru.netology.nmedia.dto.Post
 import ru.netology.nmedia.viewmodel.PostViewModel
 
-
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        val viewModel: PostViewModel by viewModels()
+class FeedFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        val binding = FragmentFeedBinding.inflate(inflater, container, false)
+        val viewModel: PostViewModel by activityViewModels()
         val adapter = PostAdapter(object : OnInteractionListener {
 
             override fun like(post: Post) {
                 viewModel.likeById(post.id)
             }
+
             override fun remove(post: Post) {
                 viewModel.removeById(post.id)
             }
@@ -36,6 +39,7 @@ class MainActivity : AppCompatActivity() {
                 viewModel.changeContent(result)
                 viewModel.save()
             }
+
             override fun edit(post: Post) {
                 editPostLauncher.launch(post.content)
                 viewModel.edit(post)
@@ -47,7 +51,8 @@ class MainActivity : AppCompatActivity() {
                     putExtra(Intent.EXTRA_TEXT, post.content)
                     type = "text/plain"
                 }
-                val shareIntent = Intent.createChooser(intent, getString(R.string.chooser_share_post))
+                val shareIntent =
+                    Intent.createChooser(intent, getString(R.string.chooser_share_post))
                 startActivity(shareIntent)
                 viewModel.shareCounter(post.id)
             }
@@ -61,15 +66,11 @@ class MainActivity : AppCompatActivity() {
 
                 val playVideo = Intent.createChooser(intent, "play Video")
                 startActivity(playVideo)
- //               Toast.makeText(this@MainActivity, "video play", Toast.LENGTH_SHORT).show()
             }
         })
 
-
-
-
         binding.recyclerView.adapter = adapter
-        viewModel.data.observe(this) { posts ->
+        viewModel.data.observe(viewLifecycleOwner) { posts ->
             val newPost = posts.size > adapter.currentList.size
             adapter.submitList(posts) {
                 if (newPost) {
@@ -78,14 +79,11 @@ class MainActivity : AppCompatActivity() {
             }
         }
 
-        val newPostLauncher = registerForActivityResult(PostResultContract()) { result ->
-            result ?: return@registerForActivityResult
-            viewModel.changeContent(result)
-            viewModel.save()
+        binding.addPostButton.setOnClickListener {
+
+            findNavController().navigate(R.id.action_feedFragment_to_newPostFragment)
         }
 
-        binding.addPostButton.setOnClickListener {
-            newPostLauncher.launch("")
-        }
+        return binding.root
     }
 }
