+++ app/src/main/res/layout/card_post.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:padding="@dimen/common_spacing"
-    tools:context=".activity.MainActivity">
+    tools:context=".screens.FeedFragment">
 
     <ImageView
         android:id="@+id/avatar"
